moved buildUrl to a module
