+++ src/clients/base-client.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const param = require('jquery-param')
+const { buildUrl } = require('../utils/url-helper')
 
 /**
  * @class Client
@@ -15,7 +15,7 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   search (parameters = {}) {
-    const url = this.__buildUrl(null, parameters)
+    const url = buildUrl(this.baseUrl, null, parameters)
     return this.sendRequest(url)
   }
 
@@ -26,7 +26,7 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   find (id, parameters = {}) {
-    const url = this.__buildUrl(id, parameters)
+    const url = buildUrl(this.baseUrl, id, parameters)
     return this.sendRequest(url)
   }
 
@@ -37,7 +37,7 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   create (data, parameters = {}) {
-    const url = this.__buildUrl(null, parameters)
+    const url = buildUrl(this.baseUrl, null, parameters)
 
     return this.sendRequest(url, {
       body: data,
@@ -53,7 +53,7 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   update (id, data, parameters = {}) {
-    const url = this.__buildUrl(id, parameters)
+    const url = buildUrl(this.baseUrl, id, parameters)
 
     return this.sendRequest(url, {
       body: data,
@@ -69,7 +69,7 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   patch (id, data, parameters = {}) {
-    const url = this.__buildUrl(id, parameters)
+    const url = buildUrl(this.baseUrl, id, parameters)
 
     return this.sendRequest(url, {
       body: data,
@@ -84,24 +84,10 @@ class BaseClient {
    * @return {promise} The request promise.
    */
   destroy (id, parameters = {}) {
-    const url = this.__buildUrl(id, parameters)
+    const url = buildUrl(this.baseUrl, id, parameters)
     return this.sendRequest(url, {
       method: 'DELETE'
     })
-  }
-
-  __buildUrl (path, parameters) {
-    let url = this.baseUrl
-    let urlPath = path || ''
-    let queryString = param(parameters)
-
-    if (url.slice(-1) !== '/') { url += '/' }
-    urlPath = urlPath.toString()
-    if (urlPath[0] === '/') { urlPath = urlPath.slice(1) }
-    url += urlPath
-    if (queryString.length > 0) { url += `?${queryString}` }
-
-    return url
   }
 }
